// src/models/kyc-error.enum.ts
export enum KycErrorEnum {
  INVALID_NIN = 'INVALID_NIN',
  NIN_NOT_FOUND = 'NIN_NOT_FOUND',
  INVALID_BVN = 'INVALID_BVN',
  BVN_NOT_FOUND = 'BVN_NOT_FOUND',
  PHONE_NOT_FOUND = 'PHONE_NOT_FOUND',
  USER_SCREENING_FAILED = 'USER_SCREENING_FAILED',
  IP_SCREENING_FAILED = 'IP_SCREENING_FAILED',
  EMAIL_CHECK_FAILED = 'EMAIL_CHECK_FAILED',
  PHONE_CHECK_FAILED = 'PHONE_CHECK_FAILED',
  INTERNAL_ERROR = 'INTERNAL_ERROR',
  KYC_LOOKUP_FAILED = 'kyc/lookup-failed',
  KYC_ALREADY_EXISTS = 'kyc/already-exists',
  NAME_MISMATCH = 'kyc/name-mismatch',
  MISSING_USER = 'MISSING_USER',
  NOT_ELIGIBLE = 'NOT_ELIGIBLE',
}