// src/models/kyc.types.ts
export enum KycProviderEnum {
  DOJAH = 'dojah',
  SUMSUB = 'sumsub',
}

export enum IdTypeEnum {
  NIN = 'NIN',
  BVN = 'BVN',
  PASSPORT = 'PASSPORT',
  DRIVER_LICENSE = 'DRIVER_LICENSE',
  VOTER_CARD = 'VOTER_CARD',
  NATIONAL_ID_CARD = 'NATIONAL_ID_CARD',
  RESIDENT_PERMIT = 'RESIDENT_PERMIT',
}